[추가] 인프라 서비스 추상화 (23.05.27 문희조)
