+++ src/main/java/com/Flaground/backend/module/member/service/MemberScheduleService.java
@@ -1,6 +1,6 @@
 package com.Flaground.backend.module.member.service;
 
-import com.Flaground.backend.infra.aws.ses.service.MailService;
+import com.Flaground.backend.infra.aws.ses.service.AwsSESServiceImpl;
 import com.Flaground.backend.module.member.domain.Member;
 import com.Flaground.backend.module.member.domain.Sleeping;
 import com.Flaground.backend.module.member.domain.repository.MemberRepository;
@@ -18,7 +18,7 @@ import java.util.stream.Collectors;
 public class MemberScheduleService {
     private final MemberRepository memberRepository;
     private final SleepingRepository sleepingRepository;
-    private final MailService mailService;
+    private final AwsSESServiceImpl mailService;
 
     //@Scheduled(cron = "000000")
     public void selectingDeactivateMembers() {
